Cleanup

+++ src/org/ojalgo/matrix/store/ShadingStore.java
@@ -29,11 +29,11 @@ package org.ojalgo.matrix.store;
  */
 abstract class ShadingStore<N extends Number> extends LogicalStore<N> {
 
-    protected ShadingStore(final MatrixStore<N> base, final long rowsCount, final long columnsCount) {
+    protected ShadingStore(final MatrixStore<N> base, final int rowsCount, final int columnsCount) {
         super(base, rowsCount, columnsCount);
     }
 
-    protected ShadingStore(final MatrixStore<N> base, final int rowsCount, final int columnsCount) {
+    protected ShadingStore(final MatrixStore<N> base, final long rowsCount, final long columnsCount) {
         super(base, rowsCount, columnsCount);
     }
